[mocks] Treat block passed to `with` as a dynamic matcher or a return value, but not both.
This fixes issue #254 in rspec-core.

The `with` method accepts two forms:

  foo.should_receive(:bar).with(:arg) { ... }
  foo.should_receive(:bar).with { ... }

In the first form, we expect foo.bar(:arg), and when it is received it
returns the value returned by eval'ing the block.

In the second form, the block itself is treated as an argument matcher.
The bug was that it was being eval'd twice because it was being treated
as both an argument matcher and a computed return value.