[core] Outputs random seed before loading files
Ordering dependencies between files can result in a raised exception
when rspec is loading files. If random ordering is turned on, then the
exception will prevent the random seed from being displayed. This
always outputs the random seed before file are loaded, so that the
ordering that produced the exception can be repeated.