[core] link to documentation for order and seed options
Since `order` and `seed` are pretty tightly coupled, it seems like
these should be documented together. But the only way I can think of
to write a feature for seed would be to try and run a spec, capture
the seed, then insert that into a config block. But that seems way too
complicated to be informative as documentation. So after consulting
with @alindeman I opted to just link to where these features are
documented for the command line options.