[core] Switch to the new `expect` syntax.
I used some vim macros to do most of the work here:
http://nathanmlong.com/2012/11/convert-to-the-new-rspec-syntax-with-vim-macros/

I also had to re-gen the html/textmate formatter output since
those include the failing line, which changed from `should`
to `expect`.

While I was at it, I deleted the files for 1.9.2-jruby because
the current jruby version supports 1.9.3, but not 1.9.2.

I wasn't able to re-gen the rbx files because my local
rbx installation is messed up and RVM is failing on me
when trying to re-install rbx :(.