[expectations] Limit diff-lcs to 1.x series only
I have some experimental code that I am working with that may change the
Diff::LCS interface in version 2.0 (whenever that drops). I would like
to encourage the adoption of the (currently bug-free) 1.2 release as
much as possible, but I don't want to cause other problems if the future
release changes.