[mocks] Remove __mock_expectation.
We want to limit the number of methods added to all object.
The use of `__send__` to get around this is a code smell
but I believe it's preferable to polluting every object with
an extra method.

This was also necessary to get the spec I added in 512bbbf7c33351332ac147feccda70400d188c60 to pass.