[core] Update caller filter.
- Don't put it at `rspec/caller_filter`.
- Thus, no need to include it in `ADDITIONAL_TOP_LEVEL_FILES`.

The way I had this was brittle and was causing problems.
In rspec-core I'm trying to update the regex and add some specs but
the fact that the file is at the same path in all 3 libs (which seemed
like a simple way to ensure only one of them loads) is causing
the _wrong_ version to be loaded in rspec-core's specs, so that it's not
picking up the changes I've made.  Putting the file at the same
path in all 3 libs was a terrible idea in retrospect :(.