[core] Move all backtrace filtering to one place
Backtrace filtering was ultimately managed by the BacktraceCleaner
(now called BacktraceFormatter), but sometimes went through what
was called the BacktraceFormatter. Specs for formatting backtraces
were in a few different places.

Now it's all managed by one object and all the specs are in one place.

Leaves a temporary delegator to support rspec-expectations until that
code is modified to use RSpec.configuration.backtrace_formatter instead
of RSpec::Core::BacktraceFormatter.

Also leaves an alias for Configuration#bactrace_cleaner. Need to decide
whether to leave that in place or deprecate it in 2.99 w/ removal in
3.0.