[core] Change dependency injection of configuration.
To me, it makes more sense to pass it to `Random.new`
and to retain it as instance state on that object.
Injecting it in the `order` method is a little odd,
as it complicates the `#order` interface for all the
other strategies, even those that don't use the
configuration at all.