[core] Swap output stream to current sandbox (Read why).
Specs that use the runner need to use the current sandbox output stream
(rather than attempting to create a new one) so that inadvertant use of
the config hasn't preloaded any formatters (e.g. deprecation warnings
will cause stuff to be initialised which would then not be captured in
our specs correctly).