[mocks] Fix issue with `d = double("foo" => 1); d.stub("foo")`.
The latter stub was being converted to a symbol, but the
first wasn't, so as a result, we wound up with two
method doubles (one for "foo", one for :foo), and when
the proxy was reset, things would blow up since it can't
reset the same method twice.

Our solution here is to centralize the `to_sym` conversion
in the `method_double` creation so that we don't have to
remember to `to_sym` the message in each place.

This allows us to remove `to_sym` from several places.
(See the last commit for the specs that give me confidence
in removing `to_sym`)