[core] Fix our simplecov setup.
Simplecov was only reporting coverage for a handful
of rspec's files, and I realized why: simplecov can
only track coverage of files that are loaded _after_
it has been loaded. So, to get this to work properly,
we have to load simplecov before rspec-core. This
commit adds a wrapper script that does that.