[mocks] Update spec to only specify the behavior we are keeping.
We're going to change the behavior of test doubles
that have been reset (so that they "expire" and can
no longer be used). This spec was originally added in
847c66d55f2ec0ffc56aa27638a3469a03ddab3d where the
point was to address an error while resetting a test
double (which, in turn caused that method double
to stick around into later examples and continue to
fail).