[mocks] Removes #to_str implememtation from test_double
Having test_double implement to_str makes the Ruby
runtime coerce it to a string whenever needed and
this isn't the intended behavior for this object.
This behavior led to issue #567 as Ruby will try
to coerce the exception object into string when raising
(here https://github.com/ruby/ruby/blob/191b373d26a48058bcc8955bf3afe426f60eaaea/eval.c#L678 )
an exception and since test_double does implement to_str
it succeded and raises that instead of failing with the
correct 'this is not an exception object' error.

This fixes #567 by providing a better error message
in it's case.