[expectations] Remove spec.
It doesn't really make sense to keep a spec that
specifies behavior when a specific gem is loaded
when there's nothing left in the project that is
custom code when that gem is loaded.

I thought it was worth committing the spec for the
history, though, so I kept it in the last commit.