[core] Remove some semantic confusion from ConfigurationOptions.
Previously it was passed a `Configuration` instance
via `configure` and also referenced `RSpec.configuration`.
In practice, these were the same instance but it was
confusing to have two references to them that could be
different.

In addition, `configure` would set `config.filter_manager`
to the local filter manager instance, which would have
led to odd behavior if the configuration instances
were different, as it could cause previously set filters
on the config instance to be discarded.