[core] Merge CommandLine into Runner.
Both contained code related to running the specs.
`CommandLine` was really just the object the
`Runner` class methods delegated to. It wasn't
related to the command line much at all -- it was
used anytime the specs were run (not just when run
from the command line) and it didn't do any command
line parsing or anything like that.

`Runner` is a better name for it. Since it has only
instance methods and the existing `Runner` has only
class methods, merging it works nicely :).

This also provides a simple, clean mechanism for alternate
runners to extend RSpec: they can subclass `Runner`
and override the class methods, the instance methods,
or either. Since we just use `self.new` to create the
runner instance, they don't have to override the
class methods in order to make it use an instance
of their subclass.