[core] Add support for the minitest + test/unit combos.
RSpec 3.x does not provide explicit support for Minitest 4.x. However,
due to differences in `require` paths, we do provide a fallback option
in the case it is used. We are not in the business of dictating which
versions our users should or should not use, nor on how they setup their
load paths (i.e. bundler or other methods).

There are a lot of subtle issues with the different combinations of
versions of Ruby + test/unit + minitest. These can cause load and name
errors.

- Update the features to specifically check for our desired assertion
  errors; instead of allowing failures due to `NameError` or `LoadError`
- Update the test/unit and minitest adapters to include fallback logic
  when different combinations occur