[mocks] Fix bug in `have_received(...).ordered`.
It worked properly when `as_null_object` or simple stubs
(via `dbl(stub_hash)` or `allow(...).to receive_messages`)
were used, but when an individual method was stubbed to allow
it to be spied on, it was registerd as one of the expectations
in the order group. When the first of multiple
`have_received(...).ordered` calls were made, it would fail
since the 2nd `have_received(...).ordered` call hadn't yet been
made but the invocation of both methods had already happened.

By ignoring stubs, it allows each `have_received` expectation
to consider only the messages that have been made part of the
order group (by the `ordered` call) so far.

Fixes #689