[core] Add an additional spec for the first from #1580.
The spec added in #1580 is a bit implicit in that
it depends upon RbConfig issuing a warning, which
may not always be the case in all versions of ruby.
The basic problem is that we were not ignoring top
level constants, so this added spec makes that
more explicit, and we can count on `Hash` always
being a top level constant even if RbConfig doesn’t
always issue the warning.