[expectations] Expand shared examples so that it shows chaining both ways.
- matcher.and other
- other.and matcher
- matcher.or other
- other.or matcher

This is in prep for supporting compound block expectations (#552).
In that case, the ordering of the matchers can matter.

However, adding these additional specs triggered a very odd
bug on 1.9.2/1.9.3 where `String#match` returned a boolean
value rather than a match data object as expected. Switching
to `Regexp#match` seems to have fixed it.