[expectations] All built-in matchers implement `expects_call_stack_jump?`.
Given the fact that our compound matcher logic calls this
(with `rescue NoMethodError`) rather than using a `respond_to?`
check, it’ll perform better to always define it.

This also changes several matcher to subclass `BaseMatcher`,
as that provides the default implementation of this new method.
It’s simpler than defining the method in the same way on all
these matchers.