[core] Remove `Example/ExampleGroup.apply?`.
We want filtering to primarily go through FilterableItemRepository
so it’s better to not even support it directly off of these classes.

Ideally, we’d refactor FilterManager to use FilterableItemRepository
but they are essentially inverted: FilterManager tracks filtering
rules, and then is asked individually about whether or not an example
(the “filterable item”) is filtered.