[core] Create only one example group alias per name.
It's possible to register the same alias multiple times. The last alias
defined, including it's metadata, will be used. This prevents the
aliases from being defined multiple times on `RSpec` and the other main
objects.

This also prevents a `NameError` when the second duplicate alias is
undefined when the DSL is no longer globally exposed.

Fix #1824