[support] Fix failing spec.
This spec was failing because of #143 and #145. Neither
PR failed when merged against the original master branch
but when both PRs were applied this began to fail because
of an interaction between the two.

Usually I would make a PR for this but I want to get our
builds back to green.