[core] Expand `metadata[:file_path]` only once.
Before we were calling `File.expand_path` and passing
`metadata[:file_path]` in many places, re-doing the same
calculation over and over again. It’s more performant
to only do that calculation once and store it in the
metadata hash for later re-use.

Note: since this now treats `:absolute_file_path` as a reserved
metadata key, this could be a breaking change for some people,
but I think it’s highly unlikely any users would use that as
a metadata key.