[mocks] Refactor odd memoization.
It is odd here because within the matcher, we call
`@expectation = expect`, and within `expect` it was
internally memoizing, which means `@expectation`
was being assigned twice. It’s also dubious to
do so because we want to minimize internal memoization
since a matcher can be re-used against a different
target.