[support] Fix intermittent test failure.
When debugging an rspec-core failure for this shared
example group I discovered that certain seeds would cause
the Tempfile to be deleted by the time we read from it.
The Ruby Tempfile docs state that it gets deleted when
the ruby Tempfile object is garbage collected.

There are ways we could work around the issue but it’s
far simpler to muse use a local non-temp file.