[core] Use "unknown" for :last_run_status when we don't have a value.
This ensures that the value is consistent — before for an unknown
case it could either be `nil` or `"unknown"`.

We've also moved the constant into configuration, so that we can
avoid the cost of loading `example_status_persister` (via constant
access since it is setup to be autoloaded) if the user hasn't
configured the feature.