[core] Remove unnecessary `uniq!` call.
No code is faster than no code.

It appears that the `uniq!` comes from micronaut:

https://github.com/rspec/rspec-core/commit/afca72820329d5f84aa60fcecb7bf188d8394919#diff-117c39fd53498f0c3f5a53efb3c24f05R52

However, it’s clear that it’s not actually needed. It didn’t do
anything for 2 years due to being a `uniq` (not `uniq!`)
call with an ignored return value, beginning in this commit:

https://github.com/rspec/rspec-core/commit/ad9281626a8e8bafe3c97d5e97a18fa9f981d74e#diff-3831ad1cf9ec21d559640a066a9d3bc8R40

…until this one, 2 years later:

67923ba940f0360740292710e798836c1c353d20

I don’t think there’s any way for there to be any duplicates to begin with, so it’s safe to remove.