[mocks] Move away from using arg splats for error generation.
We were splatting and resplatting through the error generation
process, but it was a bit confusing.  I suspect it performs
worse than just passing the same array around anyway.

One concrete improvement from this is that we had some places
where we were passing the args from a single method call to
an method accepting a splat of args_from_multiple_calls. Wrapping
the single-call in an array makes the conversion more explicit.