[expectations] Update cuke.
Limit it to demonstrate the feature using only rspec-expectations.
Using the feature in the context of rspec-core results in much
improved failure output, but will make this cuke quit brittle as
changes in rspec-core's formatting logic will break the cuke here.

I'm going to add an rspec-core cuke for this feature as well.