[core] Use :rerun_file_path when checking location filters
Fixes #1961

The problem with using :absolute_file_path is that it doesn't function
correctly for shared examples in an external file. The filter manager thinks
that there are no location filters set, when actually there are. Since
MetadataFilter already handles location filtering correctly across multiple
nested examples / files, correcting the path that is checked is sufficient to
fix the line filtering issue.